+++ src/common/configuration.h
@@ -103,6 +103,15 @@ class Configuration {
     bool editor_double_embed = false;
 
     /**
+     * If set to `true`, we'll remove the `XdndAware` property all ancestor
+     * windows in `editor.cpp`. This is needed for REAPER as REAPER implements
+     * (but doesn't use) drag-and-drop support on all of its windows. This
+     * causes the FX window to intercept the drop thus making it impossible to
+     * drag files onto plugin editors, native or otherwise.
+     */
+    bool editor_force_dnd = false;
+
+    /**
      * Use XEmbed instead of yabridge's normal editor embedding method. Wine's
      * XEmbed support is not very polished yet and tends to lead to rendering
      * issues, so this is disabled by default. Also, editor resizing won't work
@@ -173,6 +182,7 @@ class Configuration {
     void serialize(S& s) {
         s.value1b(cache_time_info);
         s.value1b(editor_double_embed);
+        s.value1b(editor_force_dnd);
         s.value1b(editor_xembed);
         s.ext(frame_rate, bitsery::ext::StdOptional(),
               [](S& s, auto& v) { s.value4b(v); });